backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
reportlab==4.0.7
django-filter==23.5
requests==2.31.0
